requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
python-docx
python-dotenv
requests
semgrep
openai
google-generativeai
